Cambiado formato
Lo mas importante es que movi de lugar el Handler, que era algo que me
estaba molestando hace rato. La razon de moverlo es porque solo puede
existir 1 EventHandler en toda la app, asi que tiene sentido que sea su
propio modulo.
