+++ src/commands/mod.rs
@@ -1,47 +1,2 @@
-use serenity::async_trait;
-use serenity::model::application::interaction::{Interaction, InteractionResponseType};
-use serenity::model::gateway::Ready;
-use serenity::model::id::GuildId;
-use serenity::prelude::*;
-
 pub mod list_users;
 pub mod ping;
-
-pub struct Handler;
-
-#[async_trait]
-impl EventHandler for Handler {
-    async fn interaction_create(&self, ctx: Context, interaction: Interaction) {
-        if let Interaction::ApplicationCommand(command) = interaction {
-            let content = match command.data.name.as_str() {
-                "ping" => ping::run(&command.data.options),
-                "users" => list_users::run(command.guild_id, &ctx, &command.data.options).await,
-                _ => "not implemented :(".to_string(),
-            };
-
-            if let Err(why) = command
-                .create_interaction_response(&ctx.http, |response| {
-                    response
-                        .kind(InteractionResponseType::ChannelMessageWithSource)
-                        .interaction_response_data(|message| message.content(content))
-                })
-                .await
-            {
-                println!("Cannot respond to slash command: {why}");
-            }
-        }
-    }
-    async fn ready(&self, ctx: Context, ready: Ready) {
-        println!("{} is connected!", ready.user.name);
-        
-        for guild in ready.guilds.iter() {
-        let commands = GuildId::set_application_commands(&guild.id, &ctx.http, |commands| {
-            commands
-                .create_application_command(|command| ping::register(command))
-                .create_application_command(|command| list_users::register(command))
-        })
-        .await;
-        println!("I now have the following guild slash commands: {commands:#?}");
-        }
-    }
-}
