+++ src/main.rs
@@ -1,4 +1,5 @@
 mod commands;
+mod handler;
 
 use dotenv::dotenv;
 use serenity::prelude::*;
@@ -12,7 +13,7 @@ async fn main() {
     let token = env::var("DISCORD_TOKEN").expect("Missing Discord Token");
 
     let mut client = Client::builder(token, GatewayIntents::empty())
-        .event_handler(commands::Handler) // Revisar commands/mod.rs
+        .event_handler(handler::Handler) // Revisar commands/mod.rs
         .await
         .expect("Error creating Client");
